Got rid of multi inheritance
