+++ src/braket/default_simulator/gate_operations.py
@@ -14,17 +14,12 @@
 from __future__ import annotations
 
 from functools import singledispatch
-from typing import Dict, List, Optional
+from typing import List
 
 import braket.ir.jaqcd as braket_instruction
 import numpy as np
-from braket.default_simulator.operation import GateOperation, Observable
-from braket.default_simulator.operation_helpers import (
-    check_hermitian,
-    check_matrix_dimensions,
-    check_unitary,
-    pauli_eigenvalues,
-)
+from braket.default_simulator.operation import GateOperation
+from braket.default_simulator.operation_helpers import check_matrix_dimensions, check_unitary
 
 
 @singledispatch
@@ -44,7 +39,7 @@ def from_braket_instruction(instruction) -> GateOperation:
     raise ValueError(f"Instruction {instruction} not recognized")
 
 
-class Identity(GateOperation, Observable):
+class Identity(GateOperation):
     """Identity gate"""
 
     def __init__(self, targets):
@@ -58,25 +53,13 @@ class Identity(GateOperation, Observable):
     def targets(self) -> List[int]:
         return self._targets
 
-    @property
-    def is_standard(self) -> bool:
-        return False
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return np.array([1, 1])
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        return None
-
 
 @from_braket_instruction.register(braket_instruction.I)
 def _i(instruction) -> Identity:
     return Identity([instruction.target])
 
 
-class Hadamard(GateOperation, Observable):
+class Hadamard(GateOperation):
     """Hadamard gate"""
 
     def __init__(self, targets):
@@ -90,29 +73,13 @@ class Hadamard(GateOperation, Observable):
     def targets(self) -> List[int]:
         return self._targets
 
-    @property
-    def is_standard(self) -> bool:
-        return True
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return pauli_eigenvalues(1)
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        # RY(-\pi / 4)
-        angle = -np.pi / 4
-        cos_component = np.cos(angle / 2)
-        sin_component = np.sin(angle / 2)
-        return np.array([[cos_component, -sin_component], [sin_component, cos_component]])
-
 
 @from_braket_instruction.register(braket_instruction.H)
 def _hadamard(instruction) -> Hadamard:
     return Hadamard([instruction.target])
 
 
-class PauliX(GateOperation, Observable):
+class PauliX(GateOperation):
     """Pauli-X gate"""
 
     def __init__(self, targets):
@@ -126,26 +93,13 @@ class PauliX(GateOperation, Observable):
     def targets(self) -> List[int]:
         return self._targets
 
-    @property
-    def is_standard(self) -> bool:
-        return True
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return pauli_eigenvalues(1)
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        # H
-        return np.array([[1, 1], [1, -1]]) / np.sqrt(2)
-
 
 @from_braket_instruction.register(braket_instruction.X)
 def _pauli_x(instruction) -> PauliX:
     return PauliX([instruction.target])
 
 
-class PauliY(GateOperation, Observable):
+class PauliY(GateOperation):
     """Pauli-Y gate"""
 
     def __init__(self, targets):
@@ -159,26 +113,13 @@ class PauliY(GateOperation, Observable):
     def targets(self) -> List[int]:
         return self._targets
 
-    @property
-    def is_standard(self) -> bool:
-        return True
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return pauli_eigenvalues(1)
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        # HS^{\dagger}
-        return np.array([[1, -1j], [1, 1j]]) / np.sqrt(2)
-
 
 @from_braket_instruction.register(braket_instruction.Y)
 def _pauli_y(instruction) -> PauliY:
     return PauliY([instruction.target])
 
 
-class PauliZ(GateOperation, Observable):
+class PauliZ(GateOperation):
     """Pauli-Z gate"""
 
     def __init__(self, targets):
@@ -191,18 +132,6 @@ class PauliZ(GateOperation, Observable):
     @property
     def targets(self) -> List[int]:
         return self._targets
-
-    @property
-    def is_standard(self) -> bool:
-        return True
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return pauli_eigenvalues(1)
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        return None
 
 
 @from_braket_instruction.register(braket_instruction.Z)
@@ -574,60 +503,3 @@ def _unitary(instruction) -> Unitary:
         return np.array([[complex(element[0], element[1]) for element in row] for row in matrix])
 
     return Unitary(instruction.targets, _from_ir_representation(instruction.matrix))
-
-
-class Hermitian(Observable):
-    """Arbitrary Hermitian observable"""
-
-    # Cache of eigenpairs for each used Hermitian matrix
-    _eigenpairs = {}
-
-    def __init__(self, targets, matrix):
-        self._targets = targets
-        clone = np.array(matrix, dtype=complex)
-        check_matrix_dimensions(clone, targets)
-        check_hermitian(clone)
-        self._matrix = clone
-
-    @property
-    def matrix(self) -> np.ndarray:
-        """np.ndarray: The Hermitian matrix defining the observable."""
-        return np.array(self._matrix)
-
-    @property
-    def targets(self) -> List[int]:
-        return self._targets
-
-    @property
-    def is_standard(self) -> bool:
-        return False
-
-    @property
-    def eigenvalues(self) -> np.ndarray:
-        return self._eigendecomposition()["eigenvalues"]
-
-    @property
-    def diagonalizing_matrix(self) -> Optional[np.ndarray]:
-        return self._eigendecomposition()["eigenvectors"].conj().T
-
-    def _eigendecomposition(self) -> Dict[str, np.ndarray]:
-        """ Decomposes the Hermitian matrix into its eigenvectors and associated eigenvalues.
-
-        The eigendecomposition is cached so that if another Hermitian observable
-        is created with the same matrix, the eigendecomposition doesn't have to
-        be recalculated.
-
-        Returns:
-            Dict[str, np.ndarray]: The keys are "eigenvectors", mapping to a matrix whose
-            columns are the eigenvectors of the matrix, and "eigenvalues", a list of
-            associated eigenvalues in the order their corresponding eigenvectors in the
-            "eigenvectors" matrix
-        """
-        mat_key = tuple(self._matrix.flatten().tolist())
-        if mat_key not in Hermitian._eigenpairs:
-            eigenvalues, eigenvectors = np.linalg.eigh(self.matrix)
-            Hermitian._eigenpairs[mat_key] = {
-                "eigenvectors": eigenvectors,
-                "eigenvalues": eigenvalues,
-            }
-        return Hermitian._eigenpairs[mat_key]
