+++ test/braket/default_simulator/test_simulation.py
@@ -13,58 +13,67 @@
 
 from collections import Counter
 
-import braket.default_simulator.operations as operation
 import numpy as np
 import pytest
+from braket.default_simulator import gate_operations, observables
 from braket.default_simulator.simulation import StateVectorSimulation
 
 evolve_testdata = [
-    ([operation.Hadamard([0])], 1, [0.70710678, 0.70710678], [0.5, 0.5]),
-    ([operation.PauliX([0])], 1, [0, 1], [0, 1]),
-    ([operation.PauliY([0])], 1, [0, 1j], [0, 1]),
-    ([operation.PauliX([0]), operation.PauliZ([0])], 1, [0, -1], [0, 1]),
-    ([operation.PauliX([0]), operation.CX([0, 1])], 2, [0, 0, 0, 1], [0, 0, 0, 1]),
-    ([operation.PauliX([0]), operation.CY([0, 1])], 2, [0, 0, 0, 1j], [0, 0, 0, 1]),
-    ([operation.PauliX([0]), operation.CZ([0, 1])], 2, [0, 0, 1, 0], [0, 0, 1, 0]),
-    ([operation.PauliX([0]), operation.Swap([0, 1])], 2, [0, 1, 0, 0], [0, 1, 0, 0]),
+    ([gate_operations.Hadamard([0])], 1, [0.70710678, 0.70710678], [0.5, 0.5]),
+    ([gate_operations.PauliX([0])], 1, [0, 1], [0, 1]),
+    ([gate_operations.PauliY([0])], 1, [0, 1j], [0, 1]),
+    ([gate_operations.PauliX([0]), gate_operations.PauliZ([0])], 1, [0, -1], [0, 1]),
+    ([gate_operations.PauliX([0]), gate_operations.CX([0, 1])], 2, [0, 0, 0, 1], [0, 0, 0, 1]),
+    ([gate_operations.PauliX([0]), gate_operations.CY([0, 1])], 2, [0, 0, 0, 1j], [0, 0, 0, 1]),
+    ([gate_operations.PauliX([0]), gate_operations.CZ([0, 1])], 2, [0, 0, 1, 0], [0, 0, 1, 0]),
+    ([gate_operations.PauliX([0]), gate_operations.Swap([0, 1])], 2, [0, 1, 0, 0], [0, 1, 0, 0]),
     (
-        [operation.PauliX([0]), operation.Swap([0, 2])],
+        [gate_operations.PauliX([0]), gate_operations.Swap([0, 2])],
         3,
         [0, 1, 0, 0, 0, 0, 0, 0],
         [0, 1, 0, 0, 0, 0, 0, 0],
     ),
-    ([operation.PauliX([0]), operation.T([0])], 1, [0, 0.70710678 + 0.70710678j], [0, 1],),
-    ([operation.PauliX([0]), operation.S([0])], 1, [0, 1j], [0, 1]),
-    ([operation.Identity([0])], 1, [1, 0], [1, 0]),
-    ([operation.Unitary([0], [[0, 1], [1, 0]])], 1, [0, 1], [0, 1]),
     (
-        [operation.PauliX([0]), operation.PhaseShift([0], 0.15)],
+        [gate_operations.PauliX([0]), gate_operations.T([0])],
+        1,
+        [0, 0.70710678 + 0.70710678j],
+        [0, 1],
+    ),
+    ([gate_operations.PauliX([0]), gate_operations.S([0])], 1, [0, 1j], [0, 1]),
+    ([gate_operations.Identity([0])], 1, [1, 0], [1, 0]),
+    ([gate_operations.Unitary([0], [[0, 1], [1, 0]])], 1, [0, 1], [0, 1]),
+    (
+        [gate_operations.PauliX([0]), gate_operations.PhaseShift([0], 0.15)],
         1,
         [0, 0.98877108 + 0.14943813j],
         [0, 1],
     ),
     (
-        [operation.PauliX([0]), operation.PauliX([1]), operation.CPhaseShift([0, 1], 0.15)],
+        [
+            gate_operations.PauliX([0]),
+            gate_operations.PauliX([1]),
+            gate_operations.CPhaseShift([0, 1], 0.15),
+        ],
         2,
         [0, 0, 0, 0.98877108 + 0.14943813j],
         [0, 0, 0, 1],
     ),
-    ([operation.RotX([0], 0.15)], 1, [0.99718882, -0.07492971j], [0.99438554, 0.00561446],),
+    ([gate_operations.RotX([0], 0.15)], 1, [0.99718882, -0.07492971j], [0.99438554, 0.00561446],),
     (
-        [operation.PauliX([0]), operation.RotY([0], 0.15)],
+        [gate_operations.PauliX([0]), gate_operations.RotY([0], 0.15)],
         1,
         [-0.07492971, 0.99718882],
         [0.00561446, 0.99438554],
     ),
-    ([operation.ZZ([0, 1], 0.15)], 2, [0.99718882 + 0.07492971j, 0, 0, 0], [1, 0, 0, 0],),
+    ([gate_operations.ZZ([0, 1], 0.15)], 2, [0.99718882 + 0.07492971j, 0, 0, 0], [1, 0, 0, 0],),
     (
-        [operation.YY([0, 1], 0.15)],
+        [gate_operations.YY([0, 1], 0.15)],
         2,
         [0.98877108, 0, 0, 0.14943813j],
         [0.97766824, 0, 0, 0.02233176],
     ),
     (
-        [operation.XX([0, 1], 0.15)],
+        [gate_operations.XX([0, 1], 0.15)],
         2,
         [0.70710678, 0, 0, -0.10566872 - 0.69916673j],
         [0.5, 0, 0, 0.5],
@@ -72,10 +81,10 @@ evolve_testdata = [
 ]
 
 apply_observables_testdata = [
-    ([operation.PauliX([0])], [operation.Hadamard([0])], 1),
+    ([observables.PauliX([0])], [gate_operations.Hadamard([0])], 1),
     (
-        [operation.PauliX([0]), operation.PauliZ([3]), operation.Hadamard([2])],
-        [operation.Hadamard([0]), operation.Identity([3]), operation.RotY([2], -np.pi / 4)],
+        [observables.PauliX([0]), observables.PauliZ([3]), observables.Hadamard([2])],
+        [gate_operations.Hadamard([0]), gate_operations.RotY([2], -np.pi / 4)],
         5,
     ),
 ]
@@ -87,9 +96,9 @@ def qft_circuit_operations():
         qft_ops = []
         for target_qubit in range(qubit_count):
             angle = np.pi / 2
-            qft_ops.append(operation.Hadamard([target_qubit]))
+            qft_ops.append(gate_operations.Hadamard([target_qubit]))
             for control_qubit in range(target_qubit + 1, qubit_count):
-                qft_ops.append(operation.CPhaseShift([control_qubit, target_qubit], angle))
+                qft_ops.append(gate_operations.CPhaseShift([control_qubit, target_qubit], angle))
                 angle /= 2
         return qft_ops
 
@@ -120,8 +129,8 @@ def test_apply_observables(observables, equivalent_gates, qubit_count):
 @pytest.mark.xfail(raises=RuntimeError)
 def test_apply_observables_fails_second_call():
     simulation = StateVectorSimulation(4)
-    simulation.apply_observables([operation.PauliX([0])])
-    simulation.apply_observables([operation.PauliX([0])])
+    simulation.apply_observables([observables.PauliX([0])])
+    simulation.apply_observables([observables.PauliX([0])])
 
 
 @pytest.mark.xfail(raises=RuntimeError)
@@ -142,7 +151,7 @@ def test_simulation_qft_circuit(qft_circuit_operations):
 
 def test_simulation_retrieve_samples():
     simulation = StateVectorSimulation(2)
-    simulation.evolve([operation.Hadamard([0]), operation.CX([0, 1])])
+    simulation.evolve([gate_operations.Hadamard([0]), gate_operations.CX([0, 1])])
     counter = Counter(simulation.retrieve_samples(10000))
     assert simulation.qubit_count == 2
     assert counter.keys() == {0, 3}
